// js/src/types/generated/zod/Target.ts
export const TargetID = withCdpMeta(z.string(), "Target.TargetID", "type");
export const SessionID = withCdpMeta(z.string(), "Target.SessionID", "type");
export const TargetInfo = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID), "type": z.string(), "title": z.string(), "url": z.string(), "attached": z.boolean(), "parentId": z.lazy(() => TargetID).optional(), "openerId": z.lazy(() => TargetID).optional(), "canAccessOpener": z.boolean(), "openerFrameId": z.lazy(() => Page.FrameId).optional(), "parentFrameId": z.lazy(() => Page.FrameId).optional(), "browserContextId": z.lazy(() => Browser.BrowserContextID).optional(), "subtype": z.string().optional() }).passthrough(), "Target.TargetInfo", "type");
// ...
export const AttachedToTargetEvent = withCdpMeta(z.object({ "sessionId": z.lazy(() => SessionID), "targetInfo": z.lazy(() => TargetInfo), "waitingForDebugger": z.boolean() }).passthrough(), "Target.attachedToTarget", "event", { phase: "event" });
export const DetachedFromTargetEvent = withCdpMeta(z.object({ "sessionId": z.lazy(() => SessionID), "targetId": z.lazy(() => TargetID).optional() }).passthrough(), "Target.detachedFromTarget", "event", { phase: "event" });
// ...
export const TargetDestroyedEvent = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID) }).passthrough(), "Target.targetDestroyed", "event", { phase: "event" });
export const TargetCrashedEvent = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID), "status": z.string(), "errorCode": z.number().int() }).passthrough(), "Target.targetCrashed", "event", { phase: "event" });
export const TargetInfoChangedEvent = withCdpMeta(z.object({ "targetInfo": z.lazy(() => TargetInfo) }).passthrough(), "Target.targetInfoChanged", "event", { phase: "event" });